all steps organized in their relevant files, at least off a quick look over for each file, ran all feature files tofether and 9 features passed 3 failed, 54 scenarios passed 8 failed, 535 steps passed 8 failed 35 skipped; it seems 2 scenarios in login and navigation, 5 in change password and 1 delete customer all fail when run together
