+++ BankingApp/Tests/Steps/DeleteCustomerSteps.py
@@ -1,6 +1,11 @@
 from behave import when
 
 # when
+@when(u'I click the Delete Information navigation button')
+def step_impl(context):
+    context.customer_poms.delete_customer_navigation().click()
+
+
 @when(u'I click the Delete Information button')
 def step_impl(context):
     context.customer_poms.delete_button().click()
