+++ BankingApp/Tests/Steps/ViewCurrentAccountsSteps.py
@@ -1,6 +1,11 @@
 from behave import when, then
 
 # when
+@when(u'I click the View Accounts navigation button')
+def step_impl(context):
+    context.account_poms.view_accounts_navigation().click()
+
+
 @when(u'I click the View Accounts button')
 def step_impl(context):
     context.account_poms.view_accounts_button().click()
