+++ BankingApp/Tests/Steps/ViewTransactionsSteps.py
@@ -1,4 +1,15 @@
-from behave import then
+from behave import when, then
+
+# when
+@when(u'I click the View Transactions button')
+def step_impl(context):
+    context.account_poms.view_transactions_button().click()
+
+
+@when(u'I click the Back button')
+def step_impl(context):
+    context.account_poms.view_transactions_back_button().click()
+
 
 # then
 @then(u'I should be on a page with the title Analyzing Your Transactions')
